Password-Strength-Meter/pyproject.toml
[project]
name = "password-strength-meter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.20",
    "langchain-community>=0.3.19",
    "langchain-core>=0.3.41",
    "langchain-google-genai>=2.0.11",
    "langgraph>=0.3.5",
    "python-dotenv>=1.0.1",
    "streamlit>=1.43.0",
]
